+++ android/common/src/main/java/org/mozilla/firefox/qt/common/Utils.kt
@@ -17,6 +17,19 @@ import java.security.KeyFactory
 import java.security.Signature
 import java.security.spec.RSAPublicKeySpec
 import java.util.*
+import android.annotation.SuppressLint
+import android.content.ContentValues
+import android.content.Context
+import android.content.Intent
+import android.net.Uri
+import android.os.Build
+import android.provider.MediaStore
+import android.provider.Settings
+import android.util.Log
+import java.io.IOException
+import java.time.LocalDateTime
+import java.time.format.DateTimeFormatter
+import java.util.Calendar
 
 // Companion for Utils.cpp
 object Utils {
@@ -65,5 +78,65 @@ object Utils {
             Intent.CATEGORY_APP_MARKET
         )
         activity.startActivity(intent)
+    }
+
+    @SuppressLint("NewApi")
+    @JvmStatic
+    fun sharePlainText(text: String): Boolean {
+        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.Q) {
+            // Not supported on oldies. :c
+            return false
+        }
+        val ctx: Context = VPNActivity.getInstance()
+        val resolver = ctx.contentResolver
+
+        // Find the right volume to use:
+        val collection = MediaStore.Downloads.getContentUri(MediaStore.VOLUME_EXTERNAL_PRIMARY)
+        val dateTime = LocalDateTime.now().format(DateTimeFormatter.ofPattern("y-mm-dd-H-m-ss"))
+        val fileMetaData =
+            ContentValues().apply {
+                put(MediaStore.Downloads.MIME_TYPE, "text/plain")
+                put(MediaStore.Downloads.DISPLAY_NAME, "MozillaVPN_Logs_$dateTime")
+                put(MediaStore.Downloads.IS_PENDING, 1)
+            }
+        // Create the File and get the URI
+        val fileURI: Uri? = resolver.insert(collection, fileMetaData)
+        if (fileURI == null) {
+            return false
+        }
+
+        val tx = resolver.openOutputStream(fileURI)
+        if (tx == null) {
+            return false
+        }
+        try {
+            val writer = tx.writer(Charsets.UTF_8)
+            writer?.write(text)
+            writer?.flush()
+        } catch (e: IOException) {
+            return false
+        }
+        tx.flush()
+        tx.close()
+        // Now update the Files meta data that the file exists
+        fileMetaData.clear()
+        fileMetaData.put(MediaStore.Downloads.IS_PENDING, 0)
+
+        try {
+            val ok = resolver.update(fileURI, fileMetaData, null, null)
+            if (ok == 0) {
+                Log.e("MozillaVPNLogs", "resolver update - err: 0 Rows updated")
+            }
+        } catch (e: Exception) {
+            Log.e("MozillaVPNLogs", "resolver update - exception: " + e.message)
+        }
+
+        val sendIntent = Intent(Intent.ACTION_SEND)
+        sendIntent.putExtra(Intent.EXTRA_STREAM, fileURI)
+        sendIntent.setType("*/*")
+
+        val chooseIntent = Intent.createChooser(sendIntent, "Share Logs")
+        ctx.startActivity(chooseIntent)
+        return true
     }
 }
