+++ android/vpnClient/src/main/java/org/mozilla/firefox/vpn/qt/VPNUtils.kt
@@ -30,66 +30,6 @@ object VPNUtils {
 
     @SuppressLint("NewApi")
     @JvmStatic
-    fun sharePlainText(text: String): Boolean {
-        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.Q) {
-            // Not supported on oldies. :c
-            return false
-        }
-        val ctx: Context = VPNActivity.getInstance()
-        val resolver = ctx.contentResolver
-
-        // Find the right volume to use:
-        val collection = MediaStore.Downloads.getContentUri(MediaStore.VOLUME_EXTERNAL_PRIMARY)
-        val dateTime = LocalDateTime.now().format(DateTimeFormatter.ofPattern("y-mm-dd-H-m-ss"))
-        val fileMetaData =
-            ContentValues().apply {
-                put(MediaStore.Downloads.MIME_TYPE, "text/plain")
-                put(MediaStore.Downloads.DISPLAY_NAME, "MozillaVPN_Logs_$dateTime")
-                put(MediaStore.Downloads.IS_PENDING, 1)
-            }
-        // Create the File and get the URI
-        val fileURI: Uri? = resolver.insert(collection, fileMetaData)
-        if (fileURI == null) {
-            return false
-        }
-
-        val tx = resolver.openOutputStream(fileURI)
-        if (tx == null) {
-            return false
-        }
-        try {
-            val writer = tx.writer(Charsets.UTF_8)
-            writer?.write(text)
-            writer?.flush()
-        } catch (e: IOException) {
-            return false
-        }
-        tx.flush()
-        tx.close()
-        // Now update the Files meta data that the file exists
-        fileMetaData.clear()
-        fileMetaData.put(MediaStore.Downloads.IS_PENDING, 0)
-
-        try {
-            val ok = resolver.update(fileURI, fileMetaData, null, null)
-            if (ok == 0) {
-                Log.e("MozillaVPNLogs", "resolver update - err: 0 Rows updated")
-            }
-        } catch (e: Exception) {
-            Log.e("MozillaVPNLogs", "resolver update - exception: " + e.message)
-        }
-
-        val sendIntent = Intent(Intent.ACTION_SEND)
-        sendIntent.putExtra(Intent.EXTRA_STREAM, fileURI)
-        sendIntent.setType("*/*")
-
-        val chooseIntent = Intent.createChooser(sendIntent, "Share Logs")
-        ctx.startActivity(chooseIntent)
-        return true
-    }
-
-    @SuppressLint("NewApi")
-    @JvmStatic
     fun openNotificationSettings() {
         val context = VPNActivity.getInstance()
         val intent = Intent()
